+++ src/RayEngine.cpp
@@ -3,8 +3,10 @@
 //
 
 #include <iostream>
-#include "include/RayEngine.h"
+
+#include "API/RayEngine.h"
 #include "Data Management/DataManagementUnit.h"
+
 
 RayEngine::RayEngine() {
     dataManagementUnit = new DataManagementUnit();
