+++ src/RayTraceEngine.pc.in
@@ -8,5 +8,5 @@ Description: @PROJECT_DESCRIPTION@
 Version: @PROJECT_VERSION@
 
 Requires:
-Libs: -L${libdir} -l@PROJECT_NAME@
+Libs: -L${libdir} -lRayTraceEngine
 Cflags: -I${includedir}
